+++ src/aws-api-mcp-server/CHANGELOG.md
@@ -11,8 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Agent script for CloudTrail Multi-Region Setup (#1320)
 - Add telemetry for AWS CLI customizations (#1335)
+- Enforcement of `AUTH_TYPE=no-auth` for streamable-http mode (#1345)
 - Agent Script for troubleshooting permissions using CloudTrail events (#1313)
-
 
 ## [0.3.0] - 2025-09-22